feat: use type erasure for partition policy in CLI
Previously, an instantiation of every sorter was compiled for each
combination of sampling and splitter sorting, leading to binary bloat and
excessive compilation times. Type erasure separates the partitioning logic
from the rest of the sorter at the cost of a single virtual function call
per invocation of a partition policy.