perf: add memory limit for radix sort in SEMS
Radix sort needs to allocate an array of bucket sizes for each depth.
In the case of the CI3 variant, this requires `(m/2)*8*2^16` bytes in
the worst case, where `m` is the longest string. With the memory limit
in place, the local sorting falls back to multi-key quicksort once an
arbitrary amount of 500MiB has been exceeded.